Added write abilities to the BLE Stream on the mobile app
The current state of this BLE stream layer of the communication stack
lacks robustness. Currently, communications with multiple devices will
result in undefined behavior. Importantly, the lock/gate values that
check to determine logical flow are shared between both reading and
writing functions, so reading and writing transmissions that happen to
occur simultaneously is currently undefined behavior. Future
implementations should split the reading and writing functionalities of
this layer and have an overall locking/sempahore to cause the process to
wait if there is a current transmission in process.