Further working
Upon subsequent testing, it appears that worker threads are not equpid
with the same scope of librariers on the main thread. As such, libraries
such as Bluetooth cannot be used. Either more research needs to be done
to find the source of the proble and fix or, or we need to settle for
work arounds in the original design of a single threaded model