Activity indicator and reobtain device settings
These changes cause the system settings list to hide and display an
activity indicator with the sign "Please Wait..." to indicate that
bluetooth transactions are in process. After the bluetooth transactions
return, the indicators are hidden, and the menu is re-displayed.

After calling reload devices, `getInitialValue()` is called to ensure
our device settings are synchronized with the system