Enable autoreconnection
These commits add a new property to the `ConnectionDelegate` class:
`reconnectAttempts` which is updated to equal the amount set in the
system settings.

A `reconnect()` function is then called if a disconnect from the app was
unintentional. In this case, the app attempts to reestablish a bluetooth
connection `reconnectAttempts` amount of times.

The `ConnectionDelegate.connect()` function was changed such that the
`bluetooth.connect()` is no longer awaited—since this was causing the
system to block and hold here. Instead we let the main thread continue,
and await on a sleep while checking if the timeout (default = 10s) has
passed. If the conenction attempt timesout, the next reconnection
attempt is started until all reconnection attempts are used up, and the
connection is determined unconnectable.