UI changes for easier access
Device Settings List has a higher height so that the settings buttons
are larger and easier to press.

Slider components within device settings menus are centered and have
less than 80% width so that the edges do not reach the edge of the
screen making it difficult to slide back if it gets to the edge.